read xyz from stdin and use kinematic engine

+++ c/kinematic_engine.h
@@ -4,6 +4,7 @@
 
 #include <math.h>
 #include <stdbool.h>
+#include <stdint.h>
 
 #define S_E_LENGTH 238
 #define E_W_LENGTH 273
@@ -18,7 +19,7 @@
 
 double law_of_cosines(double a, double b, double c);
 bool kinematic_engine(float x_pos, float y_pos, float z_pos, 
-    int * shoulder_angle, int * elbow_angle, int * wrist_angle);
+    int16_t * shoulder_angle, int16_t * elbow_angle, int16_t * wrist_angle);
 int to_deg(double radians);
